Enable updating Jinja2 by assembling size entities manually in values.
We have just a single point that prevents us updating Jinja2 to the latest version + Ansible to latest version. It's when in the group_vars we use the humanfriendly filter, which Jinja2 returns always as strings. This is a workaround of this problem and then enables https://github.com/metal-stack/metal-dockerfiles/pull/18.